pyproject.toml
[tool.poetry]
name = "bankapp"
version = "0.1.0"
description = ""
authors = ["gwy15 <gwy15thu@gmail.com>"]

[[tool.poetry.source]]
name = "tuna-pypi"
url = "https://pypi.tuna.tsinghua.edu.cn/simpl"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^1.1.2"
Flask-WTF = "^0.14.3"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.5.0"
Flask-Bcrypt = "^0.7.1"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
